Refactor how json responses are parsed in tests.
Use a map[string]interface{} type instead of a custom struct for each
type.  This saves having lots of duplicated types in the tests.

This refactoring also picked up an error in the JSON output - the
tagging on the fields of the struct wasn't correct, so the fields in the
generated JSON were CamelCased.  The tests weren't picking this up
because they had the same problem.