Rework timer event loop.
Change it to use a time.Timer so that the loop can resume listening to
the same channel.  This means that the loop will be able to handle
future actions without affecting the currently pending event.  This is
necessary to support temoprary events that will be needed for a boost
function etc.