Fix race condition in scheduler tests.
The Go 1.5 scheduler has exposed some race conditions in these tests.
Because the scheduler fires the demand function in its own goroutine,
there's a race between the state being set, and the assertins in the
tests reading it.

This replaces the global demandState variable with a dummyZone type that
encapsulates the state, and protects it with a Mutex. It also includes a
helper method to assert on the state using Gomega's asynchronous
assertion helpers.