Fix flakey thermostat test.
Depending on the order that the tests run in, the shared t variable can
be left pre-initialised from a previous test. If it's not populated with
a new instance, the Close() method will then get called for a second
time leading to a panic.  Ensuring this is set to nil at the start
prevents this happening.