Allow bulk updating sensors by deviceIDs
This will allow microcontroller sensors to be deployed without having to
have a custom config on each. They'll be able to pull the ID from their
connected sensors, and then submit reading to the new endpoint. The
config for which deviceID belongs to which sensor will then remain part
of this app.