Fix scheduler tests around DST boundaries
On the day before the start or end of DST now.AddDate(0,0,1) results in
a time that's either 23 or 25 hours ahead, not 24. This was causing the
tests to fail because they're expecting exactly 24 hours.

Most of the tests are fixed by changing the scheduler to use exactly
24hours instead of one day when there are no events. The exact duration
here doesn;t matter because it's actually an idle loop that does
nothing.

The remaining failing test was testing the wraparound at the end of the
day, so needed to calculate the expected duration based on the event
times instead of using a fixed value.