Update how ldflags are passed.
This has changed with Go 1.10. It now takes a pattern of packages to
apply the flags to. Without this, it only applies the flags to packages
explicitly listed on the commandline. When using go build without any
packages specified it doesn't apply the flags to the current package
either.

It's therefore necessary to specify the full package path when
specifying the flags.