Move event management functions up to Zone.
This reduces the number of things that need to interact with the
scheduler, allowing us to refactor the scheduler more easliy.

The zone is going to maintain its own list of events in preparation for
the scheduler being made more generic and taking a function as it's
action instead of the existing Action type.

For the moment, Boosting is left with the scheduler. This may move in
future as the shape of this evolves.