Fix some flakiness in the scheduler tests.
This wasn't asserting on the thing state after initially starting the
scheduler and waiting. This meant that occasionally the initial state
setting would happen after the boost state setting leading to the tests
failing. Adding an assertion for the thing state after starting ensures
these happen in the correct order.