Use a counterfeiter fake for gpio.Pin.
This replaces the mock from the gpio package, and removes the last usage
og gomock.

It's been necessary to use a somewhat ugly workaround to generate the
fake from the vendored gpio package due to an outstanding counterfeitter
issue[1]. I've used perl instead of sed for the substitution so that
it's portable across Mac and Linux.

[1] https://github.com/maxbrunsfeld/counterfeiter/issues/75