Add prometheus exposition of sensors and zones
In order to be able to output sensor readings along with their
corresponding timestamps we've had to implement a Collector so that we
can use ConstMetrics that allow setting the timestamp.

Zone state metrics are implemented in a more conventional way because
the values here don't come from an external system, so we are able to
report their value at now().