Fix cancelling of boost when done.
I was erroneously constructing the events using the demand function
directly instead of the trigger function, which meant that if an event
had been edited or removed, the boost state wouldn't be cleared when it
was complete.

Moving buildSchedulerJob to a method on eventHandler means there's only
one place that knows which trigger function to use, which should prevent
this inconsistency in future.