Rename Sensors deviceID to just ID.
in this context DeviceID is a confusing term because device can refer to
both the unit that's reading the temperatures and sending them in, or an
individual temperature sensor. These IDs refer to the actual temperature
sensors themselves, so in the context of a Sensor, it makes sense to
just call them ID as it's clear what they're the ID of.