Update tools for Go modules
Use the tools/tools/go pattern for ensuring that these tools are
included in the go.mod, and update the go:generate lines to call these
with `go run` so that it picks up the versions from go.mod.

Because the versions of these are now being tracked, I've pulled in the
latest versions, so the generated output for them has changed a little,
but this hasn't significantly changed the behaviour.